Separated CircuitPlugin .jj parsers
