+++ CircuitPlugin/javacc/org/workcraft/plugins/circuit/jj/verilog/VerilogParser.jj
@@ -12,7 +12,7 @@ options {
 }
 
 PARSER_BEGIN(VerilogParser)
-package org.workcraft.plugins.circuit.jj;
+package org.workcraft.plugins.circuit.jj.verilog;
 
 import java.util.List;
 import java.util.LinkedList;
