+++ CircuitPlugin/src/org/workcraft/plugins/circuit/serialisation/VerilogSerialiser.java
@@ -41,8 +41,8 @@ import org.workcraft.plugins.circuit.CircuitSettings;
 import org.workcraft.plugins.circuit.CircuitUtils;
 import org.workcraft.plugins.circuit.Contact;
 import org.workcraft.plugins.circuit.FunctionComponent;
-import org.workcraft.plugins.circuit.jj.ParseException;
-import org.workcraft.plugins.circuit.jj.SubstitutionParser;
+import org.workcraft.plugins.circuit.jj.substitution.ParseException;
+import org.workcraft.plugins.circuit.jj.substitution.SubstitutionParser;
 import org.workcraft.plugins.circuit.verilog.SubstitutionRule;
 import org.workcraft.plugins.shared.CommonDebugSettings;
 import org.workcraft.serialisation.Format;
